requirements.txt
hydra-core==1.3.2
lightning==2.5.0.post0
matplotlib==3.10.1
numpy==2.2.4
omegaconf==2.3.0
pandas==2.2.3
scikit-learn==1.6.1
torch==2.6.0
wandb==0.19.8
